Glossary and user interface defined in the index.ts
